Thu, Jan 6, 2022, 5:28 PM +00:00

+++ readability/readability.c
@@ -16,8 +16,8 @@ int main(void)
     printf("%i sentences\n", count_sentences(text));
 
 
-    int per_hundred = 100 / count_words(text); //7.14
-    printf("%i\n", per_hundred);
+    float per_hundred = 100 / count_words(text); //7.14
+    printf("%f\n", per_hundred);
     int L = round(count_letters(text) / count_words(text)) * per_hundred;    //average number of letters per 100 words
     printf("%i\n", L);
     int S = round(count_sentences(text) / count_words(text)) * per_hundred;  //average number of sentences per 100 words
